# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.ErrorBody'
      meta: {}
      success:
        type: boolean
    type: object
  github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.ErrorBody:
    properties:
      code:
        type: string
      details:
        description: optional field-level errors
      message:
        type: string
    type: object
  github_com_Ammar022_secure-ai-chat-backend_internal_subscription_dto.CreateSubscriptionRequest:
    properties:
      auto_renew:
        description: AutoRenew controls whether the subscription automatically renews
          at the end of the billing cycle.
        example: true
        type: boolean
      billing_cycle:
        description: |-
          BillingCycle must be one of: "monthly", "yearly"
          Yearly billing applies a 20% discount on the monthly price.
        enum:
        - monthly
        - yearly
        example: monthly
        type: string
      tier:
        description: |-
          Tier must be one of: "basic", "pro", "enterprise"
          basic = 10 msgs | pro = 100 msgs | enterprise = unlimited
        enum:
        - basic
        - pro
        - enterprise
        example: pro
        type: string
    required:
    - billing_cycle
    - tier
    type: object
  github_com_Ammar022_secure-ai-chat-backend_internal_subscription_dto.ToggleAutoRenewRequest:
    properties:
      enable:
        description: Enable sets auto_renew to true or false.
        example: true
        type: boolean
    type: object
...
  internal_subscription_controller.subscriptionResponse:
    properties:
      auto_renew:
        example: true
        type: boolean
      billing_cycle:
        enum:
        - monthly
        - yearly
        example: monthly
        type: string
      cancelled_at:
        example: "2024-03-15T08:00:00Z"
        type: string
      created_at:
        example: "2024-03-01T00:00:00Z"
        type: string
      end_date:
        example: "2024-04-01T00:00:00Z"
        type: string
      id:
        example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        type: string
      is_active:
        example: true
        type: boolean
      max_messages:
        example: 100
        type: integer
      messages_used:
        example: 23
        type: integer
      price:
        example: 49.99
        type: number
      remaining_messages:
        example: 77
        type: integer
      renewal_date:
        example: "2024-04-01T00:00:00Z"
        type: string
      start_date:
        example: "2024-03-01T00:00:00Z"
        type: string
      tier:
        enum:
        - basic
        - pro
        - enterprise
        example: pro
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Production-grade AI chat backend with Auth0 authentication, DDD, quota
    management, and subscription bundles.
  title: Secure AI Chat Backend
  version: "1.0"
paths:
...
  /api/v1/subscriptions:
    get:
      description: Returns all subscription bundles (active and inactive) for the
        authenticated user.
      parameters:
      - description: Unix timestamp in seconds (e.g. 1711234567)
        in: header
        name: X-Request-Timestamp
        required: true
        type: string
      - description: Unique string per request, e.g. a UUID
        in: header
        name: X-Nonce
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of subscriptions
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_subscription_controller.subscriptionResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Creates a new active subscription for the user. Users may hold
        multiple active bundles simultaneously. Quota is deducted from the newest
        bundle that still has capacity.
      parameters:
      - description: Unix timestamp in seconds (e.g. 1711234567)
        in: header
        name: X-Request-Timestamp
        required: true
        type: string
      - description: Unique string per request, e.g. a UUID
        in: header
        name: X-Nonce
        required: true
        type: string
      - description: Subscription details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_subscription_dto.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Subscription created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_subscription_controller.subscriptionResponse'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "422":
          description: Validation error — see details for field messages
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: Create a subscription bundle
      tags:
      - subscriptions
  /api/v1/subscriptions/{id}:
    get:
      description: Returns a single subscription bundle by ID. Users can only access
        their own subscriptions.
      parameters:
      - description: Unix timestamp in seconds (e.g. 1711234567)
        in: header
        name: X-Request-Timestamp
        required: true
        type: string
      - description: Unique string per request, e.g. a UUID
        in: header
        name: X-Nonce
        required: true
        type: string
      - description: Subscription UUID (e.g. a1b2c3d4-e5f6-7890-abcd-ef1234567890)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Subscription
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_subscription_controller.subscriptionResponse'
              type: object
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get a subscription
      tags:
      - subscriptions
  /api/v1/subscriptions/{id}/auto-renew:
    patch:
      consumes:
      - application/json
      description: Enables or disables automatic renewal for a subscription. Cannot
        be re-enabled on a cancelled subscription.
      parameters:
      - description: Unix timestamp in seconds (e.g. 1711234567)
        in: header
        name: X-Request-Timestamp
        required: true
        type: string
      - description: Unique string per request, e.g. a UUID
        in: header
        name: X-Nonce
        required: true
        type: string
      - description: Subscription UUID (e.g. a1b2c3d4-e5f6-7890-abcd-ef1234567890)
        in: path
        name: id
        required: true
        type: string
      - description: Set enable to true to turn on auto-renew, false to turn it off
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_subscription_dto.ToggleAutoRenewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated subscription
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_subscription_controller.subscriptionResponse'
              type: object
        "400":
          description: Invalid input or subscription is cancelled
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: Toggle auto-renew
      tags:
      - subscriptions
  /api/v1/subscriptions/{id}/cancel:
    patch:
      description: Marks the subscription as cancelled. Access is preserved until
        the end of the billing cycle. Auto-renew is disabled. Historical usage data
        is preserved.
      parameters:
      - description: Unix timestamp in seconds (e.g. 1711234567)
        in: header
        name: X-Request-Timestamp
        required: true
        type: string
      - description: Unique string per request, e.g. a UUID
        in: header
        name: X-Nonce
        required: true
        type: string
      - description: Subscription UUID (e.g. a1b2c3d4-e5f6-7890-abcd-ef1234567890)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cancelled subscription
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_subscription_controller.subscriptionResponse'
              type: object
        "400":
          description: Already cancelled or invalid ID
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_Ammar022_secure-ai-chat-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: Cancel a subscription
      tags:
      - subscriptions
...
securityDefinitions:
  BearerAuth:
    description: Auth0 JWT — prefix with "Bearer ". Obtain via Auth0 dashboard (APIs
      → Test tab) or your Auth0 login flow.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"